Did A LOT of tests with templates and pointer arithmatic, im kinda comfortable with typecasting now. Also encountered weird behavior with namespaces and templates. I will use utils as a class instead of a namespace

+++ CSD2b/synthSong/codebase/sine.cpp
@@ -1,11 +1,7 @@
-//
-// Created by Ciska Vriezenga on 06/12/2022.
-//
-
 #include "sine.h"
 #include "math.h"
 
-Sine::Sine (float frequency, float samplerate)
+Sine::Sine (double frequency, double samplerate)
   : Oscillator (frequency, samplerate) {
 
   #if DEBUG
